--- Outback/src/main/java/com/human/outback/iLogin2.java
@@ -1,10 +1,0 @@
-package com.human.outback;
-
-
-
-public interface iLogin2 {
-	int loginChk2(String userid, String passcode);
-	int getUserType(String userid);
-	Member confirm_check(String userid);
-	void upLogin(String userid);
-}
