+++ Outback/src/main/java/com/human/outback/mapper/iLogin.xml
@@ -3,6 +3,13 @@
     "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="com.human.outback.iLogin">
+<select id="loginChk2" resultType="int">
+	select count(*) from member
+	 where userid = #{param1} and passcode = #{param2}
+</select>
+<select id="getUserType" resultType="int">
+	select user_type from member where userid = #{param1}
+</select>
 <select id="getLogin" resultType="com.human.outback.Member">
 	select userid, passcode, name, mobile, gender,user_type,login_time,logout_time
 	from member
